fix 通过比特率选择清晰度v2

+++ lib/core/site/douyin_site.dart
@@ -356,7 +356,7 @@ class DouyinSite extends LiveSite with DouyinSiteMixin {
           quality: quality["name"],
           sort: level,
           data: urls,
-          bitRate: (quality["v_bit_rate"] ?? 0) / 1000,
+          bitRate: ((quality["v_bit_rate"] ?? 0) / 1000).toInt(),
         );
         if (urls.isNotEmpty) {
           for (var url in urls) {
@@ -382,7 +382,7 @@ class DouyinSite extends LiveSite with DouyinSiteMixin {
           quality: quality["name"],
           sort: quality["level"],
           data: urls,
-          bitRate: (quality["v_bit_rate"] ?? 0) / 1000,
+          bitRate: ((quality["v_bit_rate"] ?? 0) / 1000).toInt(),
         );
         if (urls.isNotEmpty) {
           for (var url in urls) {
